Refactor USB logic into abstract backend.
