+++ j5/backends/hardware/sr/v4/__init__.py
@@ -1,4 +1,11 @@
 """Backends for Student Robotics version 4 boards in the hardware environment."""
 
-from .motor_board import SRV4MotorBoardHardwareBackend  # noqa: F401
-from .power_board import SRV4PowerBoardHardwareBackend  # noqa: F401
+from .motor_board import SRV4MotorBoardHardwareBackend
+from .power_board import SRV4PowerBoardHardwareBackend
+from .servo_board import SRV4ServoBoardHardwareBackend
+
+__all__ = [
+    "SRV4MotorBoardHardwareBackend",
+    "SRV4PowerBoardHardwareBackend",
+    "SRV4ServoBoardHardwareBackend",
+]
