+++ j5/backends/hardware/sr/v4/power_board.py
@@ -2,23 +2,18 @@
 
 import struct
 from datetime import timedelta
-from functools import wraps
 from time import sleep
-from typing import (
-    Callable,
-    Dict,
-    Mapping,
-    NamedTuple,
-    Set,
-    TypeVar,
-    Union,
-    cast,
-)
+from typing import Callable, Dict, Mapping, Set, cast
 
 import usb
 
-from j5.backends import Backend, CommunicationError
 from j5.backends.hardware.env import HardwareEnvironment
+from j5.backends.hardware.j5.raw_usb import (
+    RawUSBHardwareBackend,
+    ReadCommand,
+    WriteCommand,
+    handle_usb_error,
+)
 from j5.boards import Board
 from j5.boards.sr.v4.power_board import PowerBoard, PowerOutputPosition
 from j5.components import (
@@ -28,32 +23,6 @@ from j5.components import (
     PiezoInterface,
     PowerOutputInterface,
 )
-
-
-class ReadCommand(NamedTuple):
-    """
-    Models a command to read information from the power board using USB controlRead.
-
-    code identifies the command in accordance with the definitions in usb.h in the
-    firmware source.
-
-    data_len is the number of bytes that will be returned by the command.
-    """
-
-    code: int
-    data_len: int
-
-
-class WriteCommand(NamedTuple):
-    """
-    Models a command to write information to the power board using USB controlWrite.
-
-    code identifies the command in accordance with the definitions in usb.h in the
-    firmware source.
-    """
-
-    code: int
-
 
 # The names and codes of these commands match the definitions in usb.h in the firmware
 # source.
@@ -74,38 +43,6 @@ CMD_WRITE_RUNLED = WriteCommand(6)
 CMD_WRITE_ERRORLED = WriteCommand(7)
 CMD_WRITE_PIEZO = WriteCommand(8)
 
-# Stop the library from closing the USB connections before make_safe is called.
-usb._objfinalizer._AutoFinalizedObjectBase._do_finalize_object = (  # type: ignore
-    lambda x: None
-)
-
-
-class USBCommunicationError(CommunicationError):
-    """An error occurred during USB communication."""
-
-    def __init__(self, usb_error: usb.core.USBError) -> None:
-        super().__init__(usb_error.strerror)
-
-
-RT = TypeVar('RT')
-
-
-def handle_usb_error(func: Callable[..., RT]) -> Callable[..., RT]:  # type: ignore
-    """
-    Wrap functions that use usb1 and give friendly errors.
-
-    The exceptions from usb1 are hard to find in documentation or code and are confusing
-    to users. This decorator catches the USBErrors and throws a friendlier exception that
-    can also be caught more easily.
-    """
-    @wraps(func)
-    def catch_exceptions(*args, **kwargs):  # type: ignore
-        try:
-            return func(*args, **kwargs)
-        except usb.core.USBError as e:
-            raise USBCommunicationError(e)
-    return catch_exceptions
-
 
 class SRV4PowerBoardHardwareBackend(
     PowerOutputInterface,
@@ -113,7 +50,7 @@ class SRV4PowerBoardHardwareBackend(
     ButtonInterface,
     BatterySensorInterface,
     LEDInterface,
-    Backend,
+    RawUSBHardwareBackend,
 ):
     """The hardware implementation of the SR V4 power board."""
 
@@ -151,33 +88,6 @@ class SRV4PowerBoardHardwareBackend(
         """Clean up device on destruction of object."""
         usb.util.dispose_resources(self._usb_device)
 
-    @handle_usb_error
-    def _read(self, command: ReadCommand) -> bytes:
-        return self._usb_device.ctrl_transfer(
-            0x80,
-            64,
-            wValue=0,
-            wIndex=command.code,
-            data_or_wLength=command.data_len,
-        )
-
-    @handle_usb_error
-    def _write(self, command: WriteCommand, param: Union[int, bytes]) -> None:
-        req_val: int = 0
-        req_data: bytes = b""
-        if isinstance(param, int):
-            req_val = param
-        else:
-            req_data = param
-
-        self._usb_device.ctrl_transfer(
-            0x00,
-            64,
-            wValue=req_val,
-            wIndex=command.code,
-            data_or_wLength=req_data,
-        )
-
     def check_firmware_version_supported(self) -> None:
         """Raises an exception if the firmware version is not supported."""
         v = self.firmware_version
@@ -190,13 +100,6 @@ class SRV4PowerBoardHardwareBackend(
         """The firmware version reported by the board."""
         version, = struct.unpack("<I", self._read(CMD_READ_FWVER))
         return str(cast(int, version))
-
-    @property  # type: ignore # https://github.com/python/mypy/issues/1362
-    @handle_usb_error
-    def serial(self) -> str:
-        """The serial number reported by the board."""
-        # https://github.com/python/mypy/issues/1362
-        return self._usb_device.serial_number  # type: ignore
 
     def get_power_output_enabled(self, identifier: int) -> bool:
         """Get whether a power output is enabled."""
